requirements.txt
pillow 
ffmpeg-python
tqdm

opencv-python 
tk

